+++ lib/urls.rb
@@ -1,8 +1,9 @@
 # frozen_string_literal: true
 
-require_relative 'urls/finder'
 require_relative 'urls/validator'
+require_relative 'urls/finder'
 require_relative 'urls/shortener'
+require_relative 'urls/updater'
 
 # URLs related functionality
 module URLs